:recycle: api: move Session model into enrolment context folder
