+++ api/src/certification/session/domain/usecases/validate-sessions.js
@@ -1,24 +1,24 @@
 /**
- * @typedef {import ('../../../session/domain/usecases/index.js').dependencies} deps
+ * @typedef {import ("../../../session/domain/usecases/index.js").dependencies} deps
  */
 
 import bluebird from 'bluebird';
 
 import { CertificationCandidate } from '../../../../../lib/domain/models/CertificationCandidate.js';
-import { Session } from '../models/Session.js';
+import { Session } from '../../../session/domain/models/Session.js';
 import { SessionMassImportReport } from '../models/SessionMassImportReport.js';
 
 /**
  * @param {Object} params
- * @param {deps['certificationCenterRepository']} params.certificationCenterRepository
- * @param {deps['sessionRepository']} params.sessionRepository
- * @param {deps['certificationCpfCountryRepository']} params.certificationCpfCountryRepository
- * @param {deps['certificationCpfCityRepository']} params.certificationCpfCityRepository
- * @param {deps['complementaryCertificationRepository']} params.complementaryCertificationRepository
- * @param {deps['certificationCourseRepository']} params.certificationCourseRepository
- * @param {deps['sessionCodeService']} params.sessionCodeService
- * @param {deps['sessionsImportValidationService']} params.sessionsImportValidationService
- * @param {deps['temporarySessionsStorageForMassImportService']} params.temporarySessionsStorageForMassImportService
+ * @param {deps["certificationCenterRepository"]} params.certificationCenterRepository
+ * @param {deps["sessionRepository"]} params.sessionRepository
+ * @param {deps["certificationCpfCountryRepository"]} params.certificationCpfCountryRepository
+ * @param {deps["certificationCpfCityRepository"]} params.certificationCpfCityRepository
+ * @param {deps["complementaryCertificationRepository"]} params.complementaryCertificationRepository
+ * @param {deps["certificationCourseRepository"]} params.certificationCourseRepository
+ * @param {deps["sessionCodeService"]} params.sessionCodeService
+ * @param {deps["sessionsImportValidationService"]} params.sessionsImportValidationService
+ * @param {deps["temporarySessionsStorageForMassImportService"]} params.temporarySessionsStorageForMassImportService
  */
 const validateSessions = async function ({
   sessions,
