+++ api/tests/certification/session/unit/domain/usecases/validate-sessions_test.js
@@ -1,7 +1,7 @@
 import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../../lib/domain/constants/certification-candidates-errors.js';
 import { CERTIFICATION_SESSIONS_ERRORS } from '../../../../../../lib/domain/constants/sessions-errors.js';
 import { CertificationCandidate } from '../../../../../../lib/domain/models/CertificationCandidate.js';
-import { Session } from '../../../../../../src/certification/session/domain/models/Session.js';
+import { Session } from '../../../../../../src/certification/enrollment/domain/models/Session.js';
 import { SessionMassImportReport } from '../../../../../../src/certification/session/domain/models/SessionMassImportReport.js';
 import { CpfBirthInformationValidation } from '../../../../../../src/certification/session/domain/services/certification-cpf-service.js';
 import { validateSessions } from '../../../../../../src/certification/session/domain/usecases/validate-sessions.js';
