+++ api/tests/tooling/domain-builder/factory/certification/enrolment/build-session.js
@@ -1,4 +1,4 @@
-import { Session } from '../../../../src/certification/session/domain/models/Session.js';
+import { Session } from '../../../../../../src/certification/enrolment/domain/models/Session.js';
 
 const buildSession = function ({
   id = 123,
